Wordle Crack
This is a tiny project I did which tells you the current daily wordle word and supplies an option to send you to a menu where you can put a custom date [YYYY-MM-DD] to find the word of that day. [With limits to how far in the past/future you go due to the API]

This uses the official NYT API to grab the word. 
This code is open-source and not obfuscated, you may view the contents and use it to learn, modify it, and whatnot.

Licensed under the MIT License.
You may use, modify, and distribute this code — including commercially — as long as proper credit is given.
Do not remove attribution or claim the work as your own.